deepdev/backend/requirements.txt
langgraph>=0.4.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
fastapi>=0.115.0
uvicorn>=0.34.0
websockets>=14.0
gitpython>=3.1.0
python-dotenv>=1.0.0
